User authorization

+++ app/Http/Controllers/ListingController.php
@@ -55,6 +55,12 @@ class ListingController extends Controller
 
     //Update listing data
     public function update(Request $request, Listing $listing){
+
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized action.');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -69,6 +75,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         }
 
+        $formFields['user_id'] = auth()->id;
+
         $listing->update($formFields);
 
         return back()->with('message', 'Listing updated successfully !');
@@ -76,7 +84,18 @@ class ListingController extends Controller
 
     //Delete listing
     public function destroy(Listing $listing){
+
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized action.');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully !');
+    }
+
+    //Manage listings
+    public function manage(){
+        return view('listings.manage',['listings' => auth()->user()->listings()->get()]);
     }
 }
